requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
httpx==0.28.1
pytest==8.3.4
python-dotenv==1.0.1
ruff==0.9.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
